case-8 added; fixes

+++ Kruskal.h
@@ -13,10 +13,12 @@ class ImageGraph
 
 	size_t nvertex, nedge, im_wid, im_hgt;
 
-	float calc_weigth(Vertex *n1, Vertex *n2, int im_type, bool use_distance);
+	double calc_weigth(Vertex *n1, Vertex *n2, int im_type, bool use_distance);
 	void add_edge(Vertex*, Vertex*, int im_type, bool use_distance);
 	//datastruct::Node<T>* set_vertex(T& val, int x, int y);
 	Vertex* get_vertex_by_index(int i, int j);
+	//Vertex* get_vertex_by_index24(int i, int j);
+	//Vertex* get_vertex_by_index48(int i, int j);
 public:
 	ImageGraph();
 	// constructs a graph with pixels as vertices and edge weights as the color difference
